# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fraud Scan API",
    "description": "API for fraud reporting, search, and enterprise features.\n\n**Auth methods:**\n- **Web:** JWT in `auth-token` cookie (set by login/signup).\n- **Mobile app / API clients:** See *X-API-Key + Bearer* below.\n- **External (enterprise):** Bearer token from `/enterprise/key` only.\n\n---\n\n## X-API-Key + Bearer (mobile app)\n\nEvery request must include:\n\n1. **Header: `X-API-Key`** — Shared secret key (obtain from backend team). Add to every request.\n\n2. **Header: `Authorization: Bearer <token>`** — JWT from login response. Required for protected routes (profile, search, dashboard, etc.).\n\n**Flow:**\n1. Call `POST /auth/login` with `email`, `password`, `recaptchaToken`.\n2. Read `token` from the JSON response body.\n3. Store the token and send it as `Authorization: Bearer <token>` on all protected requests.\n4. Add `X-API-Key: <your-key>` to every request.\n\n**Example headers:**\n```\nX-API-Key: <shared-key-from-backend>\nAuthorization: Bearer <jwt-from-login>\nContent-Type: application/json\n```\n\nUse the **Authorize** button above to set these and try endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://fraudscans.com/api", "description": "Production" },
    { "url": "http://localhost:3000/api", "description": "Local" }
  ],
...
  "paths": {
...
    "/fraud": {
      "post": {
        "tags": ["Fraud"],
        "summary": "Submit fraud report (cookie or guest + reCAPTCHA)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "type": { "type": "string" },
                  "severity": { "type": "string" },
                  "fraudsterDetails": { "type": "object" },
                  "evidence": { "type": "object" },
                  "guestSubmission": { "type": "object" },
                  "recaptchaToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Created, returns report id" },
          "400": { "description": "Validation" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token",
        "description": "Web: JWT in HTTP-only cookie"
      },
      "xApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Add header X-API-Key with the shared secret (get from backend team). Required on every request."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from login response body (Authorization: Bearer <token>) or enterprise API key from /enterprise/key."
      }
    }
  },
  "security": []
}